+++ packages/video_player/video_player_android/android/src/main/java/io/flutter/plugins/videoplayer/platformview/PlatformVideoView.java
@@ -21,6 +21,7 @@ import io.flutter.plugin.platform.PlatformView;
  */
 public final class PlatformVideoView implements PlatformView {
   @NonNull private final SurfaceView surfaceView;
+  @NonNull private final ExoPlayer exoPlayer;
 
   /**
    * Constructs a new PlatformVideoView.
@@ -30,7 +31,8 @@ public final class PlatformVideoView implements PlatformView {
    */
   @OptIn(markerClass = UnstableApi.class)
   public PlatformVideoView(@NonNull Context context, @NonNull ExoPlayer exoPlayer) {
-    surfaceView = new SurfaceView(context);
+    this.surfaceView = new SurfaceView(context);
+    this.exoPlayer = exoPlayer;
 
     if (Build.VERSION.SDK_INT == Build.VERSION_CODES.P) {
       // Workaround for rendering issues on Android 9 (API 28).
@@ -87,8 +89,14 @@ public final class PlatformVideoView implements PlatformView {
   }
 
   /** Disposes of the resources used by this PlatformView. */
+  @OptIn(markerClass = UnstableApi.class)
   @Override
   public void dispose() {
+    if (Build.VERSION.SDK_INT == Build.VERSION_CODES.P) {
+      exoPlayer.setVideoSurface(null);
+    } else {
+      exoPlayer.clearVideoSurfaceView(surfaceView);
+    }
     surfaceView.getHolder().getSurface().release();
   }
 }